+++ lib/l10ngen/app_localizations_ar.dart
@@ -1581,6 +1581,9 @@ class AppLocalizationsAr extends AppLocalizations {
   String get searchDateSectionTitle => 'تاريخ';
 
   @override
+  String get searchFormatSectionTitle => 'Formats';
+
+  @override
   String get searchAlbumsSectionTitle => 'الألبومات';
 
   @override