+++ lib/l10ngen/app_localizations_cs.dart
@@ -1594,6 +1594,9 @@ class AppLocalizationsCs extends AppLocalizations {
   String get searchDateSectionTitle => 'Datum';
 
   @override
+  String get searchFormatSectionTitle => 'Formats';
+
+  @override
   String get searchAlbumsSectionTitle => 'Alba';
 
   @override