+++ lib/l10ngen/app_localizations_eu.dart
@@ -1581,6 +1581,9 @@ class AppLocalizationsEu extends AppLocalizations {
   String get searchDateSectionTitle => 'Data';
 
   @override
+  String get searchFormatSectionTitle => 'Formats';
+
+  @override
   String get searchAlbumsSectionTitle => 'Albumak';
 
   @override