+++ lib/l10ngen/app_localizations_it.dart
@@ -1626,6 +1626,9 @@ class AppLocalizationsIt extends AppLocalizations {
   String get searchDateSectionTitle => 'Data';
 
   @override
+  String get searchFormatSectionTitle => 'Formats';
+
+  @override
   String get searchAlbumsSectionTitle => 'Album';
 
   @override