+++ lib/l10ngen/app_localizations_pl.dart
@@ -1591,6 +1591,9 @@ class AppLocalizationsPl extends AppLocalizations {
   String get searchDateSectionTitle => 'Data';
 
   @override
+  String get searchFormatSectionTitle => 'Formats';
+
+  @override
   String get searchAlbumsSectionTitle => 'Albumy';
 
   @override