+++ lib/l10ngen/app_localizations_pt.dart
@@ -1626,6 +1626,9 @@ class AppLocalizationsPt extends AppLocalizations {
   String get searchDateSectionTitle => 'Data';
 
   @override
+  String get searchFormatSectionTitle => 'Formats';
+
+  @override
   String get searchAlbumsSectionTitle => 'Álbuns';
 
   @override