+++ lib/l10ngen/app_localizations_ru.dart
@@ -1643,6 +1643,9 @@ class AppLocalizationsRu extends AppLocalizations {
   String get searchDateSectionTitle => 'Дата';
 
   @override
+  String get searchFormatSectionTitle => 'Formats';
+
+  @override
   String get searchAlbumsSectionTitle => 'Альбомы';
 
   @override