+++ lib/l10ngen/app_localizations_sl.dart
@@ -1629,6 +1629,9 @@ class AppLocalizationsSl extends AppLocalizations {
   String get searchDateSectionTitle => 'Date';
 
   @override
+  String get searchFormatSectionTitle => 'Formats';
+
+  @override
   String get searchAlbumsSectionTitle => 'Albums';
 
   @override