+++ lib/l10ngen/app_localizations_uk.dart
@@ -1600,6 +1600,9 @@ class AppLocalizationsUk extends AppLocalizations {
   String get searchDateSectionTitle => 'Дата';
 
   @override
+  String get searchFormatSectionTitle => 'Formats';
+
+  @override
   String get searchAlbumsSectionTitle => 'Альбоми';
 
   @override